Fixed warnings produces by MSVC
The program doesn't appear to work now but there is no warnings.
Progress?

Update:
Yes, it still works under GCC and Clang, so progress. Now I have to
debug on Windows, great :-( Let's hope it's just the test script.

+++ src/Input.cc
@@ -20,7 +20,7 @@ bool Input::refillBuffer()
 	{
 		istream.read(buffer, 1);
 		if (istream) [[likely]]
-			bufferSize = istream.readsome(buffer + 1, bufferCapacity - 1) + 1;
+			bufferSize = static_cast<std::uint16_t>(istream.readsome(buffer + 1, bufferCapacity - 1)) + 1;
 		else if (istream.eof()) [[likely]]
 			return false;
 		if (!istream) [[unlikely]]
